% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.kinesisvideomedia_package.R
\docType{package}
\name{paws.kinesisvideomedia-package}
\alias{paws.kinesisvideomedia}
\alias{paws.kinesisvideomedia-package}
\title{paws.kinesisvideomedia: Amazon Kinesis Video Streams Media}
\description{

}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker
}

}
\keyword{internal}
